# api_documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PryaniksShop API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Products": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Creates a product",
        "description": "Sample request:\r\n    \r\n    POST /api/Products\r\n    {\r\n        \"name\": \"Pryanik\",\r\n        \"description\": \"A very tasty pryanik\",\r\n        \"amount\": 60,\r\n        \"price\": 500,\r\n        \"discount\": 10\r\n    }",
        "requestBody": {
          "description": "Product to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOfPryaniks.Application.Products.Commands.CreateProduct.CreateProductCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOfPryaniks.Application.Products.Commands.CreateProduct.CreateProductCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOfPryaniks.Application.Products.Commands.CreateProduct.CreateProductCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Updates a product",
        "description": "Sample request:\r\n    \r\n    PUT /api/Products/1\r\n    {\r\n        \"id\": 1,\r\n        \"name\": \"Pryanik\",\r\n        \"description\": \"A very tasty pryanik\",\r\n        \"amount\": 60,\r\n        \"price\": 500,\r\n        \"discount\": 10\r\n    }",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An Id of item to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Product to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOfPryaniks.Application.Products.Commands.UpdateProduct.UpdateProductCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOfPryaniks.Application.Products.Commands.UpdateProduct.UpdateProductCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOfPryaniks.Application.Products.Commands.UpdateProduct.UpdateProductCommand"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "If the product Id is not the same for http address and body",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "If there is no product with this Id",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Deletes a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An Id of item to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "If there is no product with this Id",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ShopOfPryaniks.Application.Products.Commands.CreateProduct.CreateProductCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ShopOfPryaniks.Application.Products.Commands.UpdateProduct.UpdateProductCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}